\name{plot.crrQR}
\alias{plot.crrQR}
\title{
Plot estimated coefficients
}
\description{
plot method for \code{crrQR}
}
\usage{
\method{plot}{crrQR}(x, subset=NULL, main=NULL, \dots)
}
\arguments{
\item{x}{
output from \code{crrQR}
}
\item{subset }{plot a subset of coefficients}
\item{main}{main title of the plot}
\item{...}{
other arguments to plot
}}
\section{Side Effects}{
plots the variable profiles for each curve
}
\seealso{
\code{\link{crrQR}}
}
\keyword{survival}
